requirements.txt
chardet
pandas
requests
bs4
tqdm
SQLAlchemy
ollama
python-dotenv
rich